+++ web/talks/sitemaps.py
@@ -1,0 +1,26 @@
+from django.contrib import sitemaps
+from django.urls import reverse
+
+from .models import Talk
+
+
+class HomeSitemap(sitemaps.Sitemap):
+    priority = 0.6
+    changefreq = 'daily'
+
+    def items(self):
+        return ['talks:index']
+
+    def location(self, item):
+        return reverse(item)
+
+
+class TalksSitemap(sitemaps.Sitemap):
+    priority = 0.7
+    changefreq = 'daily'
+
+    def items(self):
+        return Talk.objects.all()
+
+    def lastmod(self, item):
+        return item.created
